requirements.txt
python-telegram-bot==21.*
python-dotenv==1.*
pytz==2025.*
requests==2.*
tweepy==4.*
